Add `mask-position` utilities
I also loosened the rules on position vs. origin a bit and allow classes like `*-top-left` and `*-left-top` everywhere.

The reason for this is that Tailwind used one style in some groups and another style in others in the past, but moved to support the same order in all groups while keeping the old style working. That means some class groups allow both styles and some not:
- object-position: supports both
- background-position: supports both
- mask-position: Y direction first only
- perspective-origin: Y direction first only
- transform-origin: Y direction first only

To make things simpler and keep the bundle size of tailwind-merge small, I decided to support both styles everywhere.